Closes #2496: show onboarding when fxa button is clicked
